+++ src/types/color/Color.h
@@ -12,7 +12,9 @@ class Color
 {
 public:
   Color();
-  Color(size_t r, size_t g, size_t b);
+  Color(unsigned int r,
+        unsigned int g,
+        unsigned int b);
   Color(const std::string& str, const std::string& delimiter = ",");
   
   virtual ~Color();
@@ -24,9 +26,9 @@ public:
   friend std::ostream& operator<< (std::ostream& ostr, const Color& color);
 
 protected:
-  size_t _r;
-  size_t _g;
-  size_t _b;
+  unsigned int _r;
+  unsigned int _g;
+  unsigned int _b;
 };
 
 }}}
